+++ src/widget/explore/ui/filter/ExploreFilterSection.tsx
@@ -1,8 +1,8 @@
-import { ExploreFilterList, ExploreSortingDropdown } from '@widget/explore/ui/index.ts';
+import { ExploreFilterList, ExploreSortingDropdown } from '@widget/explore/ui';
 
 const ExploreFilterSection = () => {
   return (
-    <div>
+    <div style={{ height: '100%', overflow: 'hidden' }}>
       <ExploreFilterList />
       <ExploreSortingDropdown />
     </div>
