+++ src/widget/explore/ui/tab/ExploreTabs.tsx
